Added solution for 'Diagonal Traverse'
The file contains two approaches tried:
1. 
A) Check whether the diagonal is going down or up.
b) Based on the diagonal direction and the bounds of the row and column pointers, formulate the logic for incrementing/decrementing row/column pointers.
2.
A) Start all the diagonals from either 0th row or max column index.
B) Store each diagonal array in a temporary array.
C) Check if the diagonal has to go top to down or down to up. If bottom-to-up then reverse the array before inserting that in the final array.