Added solution for 'Pascal's triangle'
The file contains two approaches I used:
1. Creates a new list for each row in the pascal triangle.
A) For the number of rows to draw in pascal's triangle, for each row create a list of integers.
B) Make the first and last element of the list to 1, always.
C) For middle elements, fetch the values of the elements in previous row at the same index and previous index, from the result list .
2.) Utilizes same list for each row in pascal's triangle.
A) Create a list of integers, that will hold the elements of the current row.
B) For each row, add a new element at 0 index with value 1 (this will act as our first element, which will be always 0)
C) Start the inner loop from 1, so first element is not checked. And for summing, consider the current element and the next element.