Added solution for 'Add Binary'
Approach:
1. Use two pointers to traverse the strings in reverse order. 
2. Use a variable to keep track of the summation of character of both the strings and any carry over.
3. Once, the sum of characters of both the strings, at specific place is calculated, we need to 
A) Make the sum either 0/1 (depending on whether the sum of two characters and carry over is 0 or 1). This can be done by using the % operator against base
B) Also, carry over needs to be modified to either 0/1 (depending on whether the previous sum of the characters was 0, 1 or 10, 11), for this we can use / operator against base 2.