Added code for 'Path Sum'
Approach:
Just start from root and check whether its value is equal to the target sum. If not, move to the left and right child and for each of them add their values to the current sum and keep doing until the leaf node is reached. If at leaf node the current sum is equal to target sum then return true else return false