Added code for 'Count Univalue Subtrees'
Approach:
[DFS - System Stack] A node is unival subtree when:
1. It is either a leaf node
2. All its children (that is left child and right child) are unival subtree itself.
So, using recursion we check whether the node has children. If not, then increment the count and return true. Then, for each left node, we do recursion and store whether that child is a unival subtree. Also, at the same time check whether the left child node value is equal to node value. Repeat the same for right child.
If at this time, if either of the left child or right child is not unival subtree, leave the count as it and return false. Else, increment the count and return true.